+++ app/src/main/java/paulscode/android/mupen64plusae/jni/CoreService.java
@@ -440,7 +440,7 @@ public class CoreService extends Service implements CoreInterface.OnFpsChangedLi
         mCoreInterface.addOnFpsChangedListener( fpsListener, fpsRecalcPeriod, mCoreInterface );
     }
 
-    void setControllerState( int controllerNum, boolean[] buttons, int axisX, int axisY, boolean isKeyboard )
+    void setControllerState( int controllerNum, boolean[] buttons, double axisX, double axisY, boolean isKeyboard )
     {
         if (!mUseRaphnetDevicesIfAvailable) {
             NativeInput.setState( controllerNum, buttons, axisX, axisY, isKeyboard );
